requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.8.0
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1695989934239/work
build==1.2.1
cachetools==5.3.3
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
chardet==5.2.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
chroma-hnswlib==0.7.3
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.5.14
Deprecated==1.2.14
entrypoints==0.4
exceptiongroup==1.2.1
faiss==1.7.4
faiss-cpu==1.7.4
fastapi==0.110.2
filelock==3.13.4
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.22.2
humanfriendly==10.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1713279365350/work
importlib-metadata==7.0.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
langchain==0.0.284
langsmith==0.0.92
libmagic==1.0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.10.0
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1707225416355/work/dist/numpy-1.26.4-cp310-cp310-macosx_11_0_arm64.whl#sha256=14c628943f67e9c10d22faa9c1cdcbf69c3192e88f2f597d03b77a252e589567
oauthlib==3.2.2
onnxruntime==1.17.3
openai==0.28.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pinecone-client==4.0.0
posthog==3.5.0
protobuf==3.20.3
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0b1
Pygments==2.17.2
Pympler==1.0.1
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
regex @ file:///Users/runner/miniforge3/conda-bld/regex_1713305278113/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
streamlit==1.22.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.4.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
unstructured==0.9.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1708239446578/work
uvicorn==0.29.0
uvloop==0.19.0
validators==0.28.1
watchfiles==0.21.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
